requirements.txt
aiodns==3.2.0
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
backports.zoneinfo==0.2.1
backtrader==1.9.78.123
binance-connector==3.8.0
blinker==1.8.2
ccxt==4.3.59
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
dateparser==1.2.0
Flask==3.0.3
frozenlist==1.4.1
idna==3.7
imaplib2==3.6
importlib_metadata==8.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
numpy==1.24.4
pandas==2.0.3
pycares==4.4.0
pycparser==2.22
pycryptodome==3.20.0
python-binance==1.0.19
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
regex==2024.5.15
requests==2.32.3
six==1.16.0
ta==0.11.0
TA-Lib==0.4.32
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.2
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
yarl==1.9.4
zipp==3.19.2
